renaming Action and removing tests for now
