+++ packages/destination-actions/src/destinations/twilio-messaging/sendMessage/__tests__/index.test.ts
@@ -1,0 +1,8 @@
+
+
+describe('TwilioMessaging.sendMessage', () => {
+  // TODO: Implement tests
+  it('should pass', () => {
+    expect(true).toBe(true)
+  })
+})
