--- packages/destination-actions/src/destinations/twilio-messaging/sendSms/__tests__/index.test.ts
@@ -1,9 +1,0 @@
-import nock from 'nock'
-import { createTestEvent, createTestIntegration } from '@segment/actions-core'
-import Destination from '../../index'
-
-const testDestination = createTestIntegration(Destination)
-
-describe('TwilioMessaging.sendSms', () => {
-  // TODO: Test your action
-})
